fix(SqlVariant): read an unsigned TINYINT faithfully instead of as signed int8_t
SQL Server's TINYINT is unsigned 0..255, but the variant read SQL_TINYINT into a
signed int8_t. A value above 127 then went one of two wrong ways: the per-row
path had the driver reject the whole row with SQLSTATE 22003 ("numeric value out
of range"), and -- because the column read sits inside a noexcept cursor -- that
rejection unwound into a hard crash; the block-prefetch path silently wrapped
255 to -1. A row was enough to hit it: any table with both a LOB column (which
disables prefetch) and a TINYINT above 127 crashed on read.

Read SQL_TINYINT into a signed short in both paths. Short holds the whole
unsigned-tinyint range, so the value round-trips and the two paths still agree on
the alternative -- the same widening the SMALLINT case already applies for the
mirror-image reason. The existing TINYINT test only used 42, below the boundary;
it now uses 200, and new cases pin 255 on its own and 255 after a LOB column.

+++ src/Lightweight/DataBinder/SqlVariant.cpp
@@ -48,7 +48,11 @@ SQLRETURN SqlDataBinder<SqlVariant>::GetColumn(
             returnCode = SqlDataBinder<bool>::GetColumn(stmt, column, &variant.emplace<bool>(), indicator, cb);
             break;
         case SQL_TINYINT:
-            returnCode = SqlDataBinder<int8_t>::GetColumn(stmt, column, &variant.emplace<int8_t>(), indicator, cb);
+            // SQL Server's TINYINT is unsigned 0..255, but SQL_C_STINYINT tops out at 127, so a value
+            // above it makes the driver report 22003 ("numeric value out of range") and fail the whole
+            // row. Read into a signed short, which holds the entire unsigned-tinyint range -- the same
+            // widening the SMALLINT case below applies for the mirror-image reason.
+            returnCode = SqlDataBinder<short>::GetColumn(stmt, column, &variant.emplace<short>(), indicator, cb);
             break;
         case SQL_SMALLINT:
             // SQL_SMALLINT is a signed 16-bit integer. Fetching into `unsigned short` makes the
